% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmReadExonTable.R
\name{ddmReadExonTable}
\alias{ddmReadExonTable}
\title{Read and process SOPHiA exon coverage table}
\usage{
ddmReadExonTable(fpath, ver = NULL)
}
\arguments{
\item{fpath}{Character, path(s) to exon coverage table}

\item{ver}{Character scalar, DDM version}
}
\value{
\link[data.table:data.table]{data.table::data.table} object with exon coverage data
}
\description{
Read and process SOPHiA exon coverage table
}
\details{
No modifications. Will add 'DDM_version' field when 'ver' parameter NOT NULL.
}
