% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmReadVariantTable.R
\name{ddmReadVariantTable}
\alias{ddmReadVariantTable}
\title{Read and process SOPHiA variant table}
\usage{
ddmReadVariantTable(fpath, ver = NULL)
}
\arguments{
\item{fpath}{Character, path(s) to variant table}

\item{ver}{Character scalar, DDM version}
}
\value{
\link[data.table:data.table]{data.table::data.table} object with variant data
}
\description{
Read and process SOPHiA variant table
}
\details{
Only modification currently, is removing '\%' from var_percent field and
coercing to numeric. Will add 'DDM_version' field when 'ver' parameter NOT
NULL.
}
