% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getTemplate.R
\name{getTemplate}
\alias{getTemplate}
\title{Get filepath for report template}
\usage{
getTemplate(name)
}
\arguments{
\item{name}{Character scalar, the report name}
}
\description{
Get filepath for report template
}
