% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tableDiffs.R
\name{tableDiffs}
\alias{tableDiffs}
\title{Summarize the differences between two tables}
\usage{
tableDiffs(
  d0,
  d1,
  mrgFlds,
  d0name = "old",
  d1name = "new",
  gcol = getOption("UNCMolGen.gcol", GOODCOL),
  bcol = getOption("UNCMolGen.gcol", BADCOL)
)
}
\arguments{
\item{d0}{\link[data.table:data.table]{data.table::data.table} object, passed to \link[arsenal:comparedf]{arsenal::comparedf} 'x'}

\item{d1}{\link[data.table:data.table]{data.table::data.table} object, passed to \link[arsenal:comparedf]{arsenal::comparedf} 'y'}

\item{mrgFlds}{Character, passed to \link[arsenal:comparedf]{arsenal::comparedf} 'by'}

\item{d0name, d1name}{Character scalar, names for the objects to appear in
tables}

\item{gcol}{Color indicating no changes}

\item{bcol}{Color indicating changes present}
}
\description{
Summarize the differences between two tables
}
