% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNCMolGen-package.R
\docType{package}
\name{UNCMolGen-package}
\alias{UNCMolGen}
\alias{UNCMolGen-package}
\title{UNCMolGen: Tools for UNC Molecular Genetics Laboratory}
\description{
Provides support for the UNC Molecular Genetics Laboratory.
}
\author{
\strong{Maintainer}: Dayne L Filer \email{dayne.filer@gmail.com} (\href{https://orcid.org/0000-0002-3443-5315}{ORCID})

}
\keyword{internal}
