% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-valDiffBoxplot.R
\name{valDiffBoxplot}
\alias{valDiffBoxplot}
\title{Boxplot of value differences from two tables}
\usage{
valDiffBoxplot(
  d0,
  d1,
  mrg,
  fld,
  grpFld = "sampleID",
  gcol = getOption("UNCMolGen.gcol", GOODCOL),
  bcol = getOption("UNCMolGen.gcol", BADCOL)
)
}
\arguments{
\item{d0}{\link[data.table:data.table]{data.table::data.table} object, the first table}

\item{d1}{\link[data.table:data.table]{data.table::data.table} object, the second table}

\item{mrg}{Character, the list of fields to merge the two data.table objects}

\item{fld}{Character scalar, the field to plot}

\item{grpFld}{Character scalar, the field to group by}

\item{gcol}{Color for boxes WITHOUT changes in fld from d0 to d1}

\item{bcol}{Color for boxes WITH changes in fld from d0 to d1}
}
\description{
Draw a boxplot of value differences from the merge of two tables
}
